move the rendering utils into the map
